Update deprecated vscode package dependency to vscode-test
These changes replace the deprecated "vscode" package dependency with
"vscode-test", which fixed the majority of the security vulnerabilities
reported by npm. This dependency is now properly listed as a dependency,
as opposed to a development dependency.

In addition, the typescript-related dependencies have been removed since
they were unused in this project. This should decrease the chance of
further vulnerabilities as well as reduce the bundled extension size.